Move shape render util back to renderNG
Move shape render util back to renderNG
